+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 2025-05-30 02:00:02
+### Changed
+- Updated `publish_new_release.ps1` to auto-detect `PackageIdentifier` if not provided.
+- Updated `publish_new_release.ps1` to include itself in the files staged for the release commit.
 ## [0.1.5] - 2025-05-30 01:58:21
 ### Changed
 - Investigated 401 Unauthorized errors for Gen1 Miniservers during version checks.